Split SDL project up into more managable pieces (#473)
* Enable text for SDL app on mac

* fix screenshots

* brew readme

* wip

* Split SDL project up into more managable pieces

* clean

* clean

* Fixed Windows build, made (unused) FPS counter portable to non-POSIX platforms.
