+++ src/libprojectM/projectM-opengl.h
@@ -11,7 +11,7 @@
 // Enable openGL extra checks, better not be enabled in release build
 #define OGL_DEBUG                   0
 
-// Unlock FPS for rendering benchmarks, it disables Vblank/Vsync and prints drawned frames count within a 5s test run
+// Unlock FPS for rendering benchmarks, it disables Vblank/Vsync and prints drawn frame count within a 5s test run
 #define UNLOCK_FPS                  0
 
 // If a shader compilation failure occurs, it dumps shader source into /tmp instead of stderr
