+++ src/projectM-sdl/audioCapture.hpp
@@ -1,0 +1,6 @@
+#ifndef audioCapture_hpp
+#define audioCapture_hpp
+
+#include "pmSDL.hpp"
+
+#endif /* audioCapture_hpp */
