+++ src/projectM-sdl/loopback.hpp
@@ -1,0 +1,14 @@
+
+
+#ifndef loopback_hpp
+#define loopback_hpp
+
+#include "pmSDL.hpp"
+
+class projectMSDL;
+
+bool initLoopback();
+void configureLoopback(projectMSDL *app);
+bool processLoopbackFrame(projectMSDL *app);
+
+#endif /* loopback_hpp */
